Fixes issue #62, samples at proper gait cycle percentages.
The gait cycles extracted by split_at no longer show data from heel strike to
heel strike, for example. The start and end of cycle times are found and the
data is interpolated at the desired number of samples, with the last sample
time being cycle_duration * (1.0 - 1.0 / num_desired_samples). This was
incorrectly including the start cycle time from the following cycle in the
previous cycle.