Adds the ability to remap marker labels from the meta.yml file.
This fixes issue #68. It adds a new field in the meta data in which the user
can provide a mappign between the marker names in the raw data files and the
desired marker names in the resulting data frames returned by DFlowData
methods. It follows the same design pattern as the analog channel relabeling
method.