Adds the ability to deal with missing markers post D-Flow 3.16.1.
As expliained in issue #69, D-Flow's output of missing markers in 3.16.2rc4
changed to strings of zeros instead of holding the value constant. So the
toolkit must deal with pre and post variations on the missing markers because
we've collected data in both situations. This commit disables the previous
missing marker identifier if a D-Flow version > 3.16.1 is provided in the meta
data file. If it isn't provided it defaults to the current missing marker
identification behavior based on strings of zeros. Note that this does not yet
properly deal with fixing missing markers in the compensation trials because we
do not currently load in the meta data for the compensation trials. This will
need to be fixed once we have to use the compensation trials for lateral and
pitching motions. The ultimate solution for the compensation trials may be to
subclass DFlowData specifically for compensation data and put all the methods
associated with it in that class. Then the marker identification would come
from the cues in the compensation meta data.

This commit also introduces some slight backwards incompatible changes to
_load_mocap_data and clean_data (which is public). These slight changes to the
arguments and the defaults allow for the missing values to be fixed in both the
marker and HBM data. Previously, only the marker's were interpolated. This
defaults for clean_data are more intuitive, in my opinion.