Fixed installation issues and bumped to 0.1.1.
I did not have the package_data specified in the setup.py file so none of the
octave files and none of the data files are included in the installation
directories (they were most included in a source distribution, but not
installed). This caused pip installs to be broken, for example.

Version is bumped to 0.1.1 as a bug fix release and will be merge into the
0.2.0dev in master.