Speeds up the data frame interpolation function.
The gait/interpolate function is only used in the GaitData.split_at() method
and the way I orginally coded it using Pandas paradigms was very slow. This
change speeds things up by about 40X. I had to change the test because it no
longer supports nans in the DataFrame. Being that it is only used in the
split_at method this seemed ok. So far I've never had nans in the data frame by
the time it gets to the split at method. GaitData isn't built to handle nans,
they should be fixed before creating a GaitData object.